Make timers work with old tokio
tokio_timer 0.2 requires a `Timer` that is constantly polled in order to
function. If no such `Timer` exists, all the timer-related methods will
fail. The intention is that this is taken care of by the new tokio
runtime, but in our tsunami setup, we're still using the *old* tokio,
and as such there isn't a runtime running!

So, for now, we revert to the old (and fairly clumsy) tokio_timer 0.1.