Allow setting EC2 availability zone
This adds `set_availability_zone` on `TsunamiBuilder`, which callers can
use to set the EC2 availability zone (which is more fine-grained than an
EC2 "region") for the spot instances to be launched.

Note that the availability zone is passed as a string, and no validation
against the set `Region` takes place, so it is currently possible to set
a mismatching region and AZ (e.g., `Region::UsEast1` vs. AZ
us-west-1a").